split Core into Neutral and AspNetCore
