+++ Sekiban.sln
@@ -95,6 +95,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pure.Domain.Test", "tests\P
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sekiban.Core", "src\Sekiban.Core\Sekiban.Core.csproj", "{524AE6A7-5B7C-4D22-B701-F926710BBBAA}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sekiban.Core.AspNetCore", "src\Sekiban.Core.AspNetCore\Sekiban.Core.AspNetCore.csproj", "{8DA8CCA4-3A14-4626-8883-207E80906911}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -269,6 +271,10 @@ Global
 		{524AE6A7-5B7C-4D22-B701-F926710BBBAA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{524AE6A7-5B7C-4D22-B701-F926710BBBAA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{524AE6A7-5B7C-4D22-B701-F926710BBBAA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8DA8CCA4-3A14-4626-8883-207E80906911}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8DA8CCA4-3A14-4626-8883-207E80906911}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8DA8CCA4-3A14-4626-8883-207E80906911}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8DA8CCA4-3A14-4626-8883-207E80906911}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -316,6 +322,7 @@ Global
 		{59FE4616-6610-48AE-A44A-2AC2E1D27A05} = {B85F80AE-6A27-402C-93FA-352D938EDA8D}
 		{51DF08A1-D72A-476E-80EC-77257CAD3132} = {4C34F748-F223-44DD-899D-029F9CEF7780}
 		{524AE6A7-5B7C-4D22-B701-F926710BBBAA} = {76D7B2DA-2E6B-4078-AC01-B29BCA83411A}
+		{8DA8CCA4-3A14-4626-8883-207E80906911} = {76D7B2DA-2E6B-4078-AC01-B29BCA83411A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2DFBE53E-69A7-453A-8D20-72271AE32016}
