+++ internalUsages/FeatureCheck.WebApi/Program.cs
@@ -2,9 +2,9 @@ using FeatureCheck.Domain.Aggregates.Branches.Commands;
 using FeatureCheck.Domain.Shared;
 using FeatureCheck.Domain.Usecases;
 using Microsoft.Azure.Cosmos;
-using Sekiban.Core.Command;
+using Sekiban.Core.AspNetCore.Command;
+using Sekiban.Core.AspNetCore.Usecase;
 using Sekiban.Core.Dependency;
-using Sekiban.Core.Usecase;
 using Sekiban.Infrastructure.Cosmos;
 using Sekiban.Infrastructure.Cosmos.Lib.Json;
 using Sekiban.Web.Dependency;
