+++ src/Sekiban.Core.AspNetCore/Sekiban.Core.AspNetCore.csproj
@@ -7,8 +7,11 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Sekiban.Core.Neutral\Sekiban.Core.Neutral.csproj" />
-        <ProjectReference Include="..\Sekiban.Core.AspNetCore\Sekiban.Core.AspNetCore.csproj" />
+        <FrameworkReference Include="Microsoft.AspNetCore.App"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\Sekiban.Core.Neutral\Sekiban.Core.Neutral.csproj"/>
     </ItemGroup>
 
 </Project>
