+++ src/Sekiban.Core.Neutral/Sekiban.Core.Neutral.csproj
@@ -23,7 +23,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <FrameworkReference Include="Microsoft.AspNetCore.App"/>
         <PackageReference Include="MediatR" Version="12.4.1"/>
         <PackageReference Include="MemStat.Net" Version="0.1.4"/>
         <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="9.0.0"/>
