--- src/Sekiban.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Sekiban.Core;
-
-public class Class1
-{
-
-}
